requirements.txt
django==3.2.12
djangorestframework==3.13.1
django-cors-headers